docker-compose: run container in host pid namespace
By default, containers run in a separate namespace. This isn't very
interesting for htop, because we can't see any processes that aren't in
our namespace.

We could probably also add an option to set the namespace to be shared
with another container, in case we don't want to run containers in the
host pid namespace.